pyproject.toml
[project]
name = "langchain-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "boto3>=1.40.17",
    "flask>=3.1.2",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-aws>=0.2.31",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.7",
    "langchain-tavily>=0.2.11",
    "langchainhub>=0.1.21",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
